@@ -1,6 +1,6 @@
 {
   "name": "default",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -10,21 +10,23 @@
     "watch": "vue-cli-service build --watch --mode development"
   },
   "dependencies": {
-    "@vuepic/vue-datepicker": "^3.3.1",
-    "axios": "^0.27.2",
-    "core-js": "^3.8.3",
-    "vue": "^3.2.13",
-    "vue-axios": "^3.4.1",
-    "vuedraggable": "^4.1.0"
+    "@vuepic/vue-datepicker": "^3.5.2",
+    "axios": "^1.1.3",
+    "core-js": "^3.25.5",
+    "dayjs": "^1.11.6",
+    "node-polyfill-webpack-plugin": "^2.0.1",
+    "vue": "^3.2.41",
+    "vue-axios": "^3.5.0",
+    "vue3-timepicker": "^1.0.0-beta.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3"
+    "@babel/core": "^7.19.6",
+    "@babel/eslint-parser": "^7.19.1",
+    "@vue/cli-plugin-babel": "~5.0.8",
+    "@vue/cli-plugin-eslint": "~5.0.8",
+    "@vue/cli-service": "~5.0.8",
+    "eslint": "^8.26.0",
+    "eslint-plugin-vue": "^9.6.0"
   },
   "eslintConfig": {
     "root": true,
@@ -39,7 +41,9 @@
       "parser": "@babel/eslint-parser"
     },
     "rules": {
-      "vue/multi-word-component-names": "off"
+      "vue/multi-word-component-names": "off",
+      "no-console": "off",
+      "no-unused-vars": "off"
     }
   },
   "browserslist": [
