@@ -14,7 +14,7 @@ class extKalenderAdminList extends AbstractPage {
 	public function execute() {
 
 		//$this->getRequest();
-		//$this->getAcl();
+		$acl = $this->getAcl();
 
         $user = DB::getSession()->getUser();
 
@@ -25,10 +25,12 @@ class extKalenderAdminList extends AbstractPage {
         $this->render([
             "tmplHTML" => '<div class="box"><div class="box-body"><div id=app></div></div></div>',
             "scripts" => [
-                PATH_EXTENSION.'tmpl/scripts/list/dist/main.js'
+                PATH_EXTENSION . 'tmpl/scripts/list/dist/js/chunk-vendors.js',
+                PATH_EXTENSION . 'tmpl/scripts/list/dist/js/app.js'
             ],
             "data" => [
-                "apiURL" => "rest.php/kalender"
+                "apiURL" => "rest.php/kalender",
+                "acl" => $acl['rights']
             ]
 
         ]);
