@@ -72,6 +72,7 @@ class extKalenderModelKalender
     }
 
 
+
     /**
      * Collection
      */
@@ -85,12 +86,17 @@ class extKalenderModelKalender
             "color" => $this->getColor(),
             "sort" => $this->getSort(),
             "preSelect" => $this->getPreSelect(),
-            "acl" => $this->getAcl(),
+            "aclID" => $this->getAcl(),
             "ferien" => $this->getFerien(),
-            "public" => $this->getPublic(),
+            "public" => $this->getPublic()
         ];
         if ($full == true) {
 
+            if ($collection['aclID']) {
+                $collection['acl'] = ACL::getAcl( DB::getSession()->getUser(), false, (int)$collection['aclID'] );
+            } else {
+                $collection['acl'] = ACL::getBlank();
+            }
 
         }
 
@@ -102,19 +108,170 @@ class extKalenderModelKalender
     /**
      * @return Array[]
      */
-    public static function getAll() {
+    public static function getAll($state = false) {
 
+        /*
+        $where = '';
+        if ($state) {
+            $where = 'WHERE state = '.(int)$state;
+        }
         $ret =  [];
         $dataSQL = DB::getDB()->query("SELECT  a.*
             FROM ext_kalender as a
-            WHERE state = 1
+            $where
             ORDER BY a.sort ");
         while ($data = DB::getDB()->fetch_array($dataSQL, true)) {
             $ret[] = new self($data);
         }
         return $ret;
+        */
+
+        $ret =  [];
+        if ($state) {
+            $data = DB::run( 'SELECT * FROM ext_kalender WHERE state = :state ORDER BY sort', ["state" => (int)$state] )->fetchAll();
+        } else {
+            $data = DB::run( 'SELECT * FROM ext_kalender ORDER BY sort' )->fetchAll();
+        }
+
+        foreach($data as $item) {
+            $ret[] = new self($item);
+        }
+        return $ret;
     }
 
+
+
+    public static function updateSort($items)
+    {
+
+        if (!$items || count($items) < 1) {
+            return false;
+        }
+
+        foreach ($items as $item) {
+
+            if ( !DB::run( 'UPDATE ext_kalender SET sort = :state WHERE id = :id ',
+                ["id" => (int)$item->id, "state" => (int)$item->sort] )) {
+                return false;
+            }
+
+        }
+        return true;
+
+    }
+
+    public static function updateAcl($id, $acl)
+    {
+        if ( DB::run( 'UPDATE ext_kalender SET acl = :acl WHERE id = :id ',
+            ["id" => (int)$id, "acl" => (int)$acl] ) ) {
+            return true;
+        }
+        return false;
+    }
+
+    public static function updateState($id, $state)
+    {
+        if ( DB::run( 'UPDATE ext_kalender SET state = :state WHERE id = :id ',
+            ["id" => (int)$id, "state" => (int)$state] ) ) {
+            return true;
+        }
+        return false;
+    }
+
+    public static function submitData($array) {
+
+        if (!$array) {
+            return false;
+        }
+        if (!$array['title']) {
+            return false;
+        }
+
+        if ( !$array['state'] || (int)$array['state'] == 0) {
+            $array['state'] = 0;
+        }
+        if ( !$array['color'] || $array['color'] == 'undefined' ) {
+            $array['color'] = '';
+        }
+
+        $insert_id = 0;
+        if ($array['id']) {
+
+            // UPDATE !
+            if (!DB::getDB()->query("UPDATE ext_kalender SET
+                        title = '" . DB::getDB()->escapeString($array['title']) . "',
+                        state = " . (int)DB::getDB()->escapeString($array['state']) . ",
+                        color = '" . DB::getDB()->escapeString($array['color']) . "',
+                        sort = " . (int)DB::getDB()->escapeString($array['sort']) . ",
+                        preSelect = " . (int)DB::getDB()->escapeString($array['preSelect']) . ",
+                        acl = " . (int)DB::getDB()->escapeString($array['acl']) . ",
+                        ferien = " . (int)DB::getDB()->escapeString($array['ferien']) . ",
+                        public = " . (int)DB::getDB()->escapeString($array['public']) . "
+                        WHERE id = " . (int)$array['id'])) {
+                return false;
+            }
+            $insert_id = (int)$array['id'];
+
+        } else {
+
+            // INSERT !
+            if ( !DB::getDB()->query("INSERT INTO ext_kalender
+            (
+                state,
+                title,
+                color,
+                sort,
+                preSelect,
+                acl,
+                ferien,
+                public
+            ) values(
+            1,
+            '" .  DB::getDB()->escapeString($array['title']) . "',
+            '" .  DB::getDB()->escapeString($array['color']) . "',
+            " .  (int)DB::getDB()->escapeString($array['sort']) . ",
+            " . (int)DB::getDB()->escapeString($array['preSelect']) . ",
+            " . (int)DB::getDB()->escapeString($array['acl']) . ",
+            " . (int)DB::getDB()->escapeString($array['ferien']) . ",
+            " . (int)DB::getDB()->escapeString($array['public']) . "
+            ) ") ) {
+                return false;
+            }
+            $insert_id = DB::getDB()->insert_id();
+        }
+
+
+
+        return $insert_id;
+    }
+
+    public static function deleteFromID( $id ) {
+
+        if (!$id) {
+            return false;
+        }
+
+        if (!DB::getDB()->query("DELETE FROM ext_kalender WHERE id=".(int)$id)) {
+            return false;
+        }
+        return true;
+
+    }
+
+    public static function deleteALL(  ) {
+
+        if (!DB::getDB()->query("TRUNCATE TABLE ext_kalender")) {
+            return false;
+        }
+        return true;
+    }
+
+    public static function countAll(  ) {
+        if ($data = DB::getDB()->query_first("SELECT COUNT(id) AS count FROM ext_kalender; ")) {
+            return $data['count'];
+        }
+        return true;
+    }
 
 
 
